Added IncludeSysSchema switch, corrected query per recommendations
Added the IncludeSysSchema switch to allow user to control the display
of sys schema objects.  Corrected the main SQL query as requested.
Added CONVERTs to the ABS CASE statements to avoid an arithmetic
overflow condition.  After analyzing the possible columns to sort by, I
decided PercentUsed is the most useful and I would like to keep that as
the default output.